Initial commit
Added three demos from Google's MediaPipe API for Face, Hand and Pose landmark detection.

Adjusted the Hand landmark detection, so you can log frames of data, instead of constantly doing it.

Added a function to convert raw landmark data to the correct format for training. I implemented this function, so it converts data directly when logging landmarks (in the hand demo), so you don't have to do it manually.

Trained a model to detect neutral hand positions and a pointing hand. The data was saved as JSON objects. I used an extra hidden layer for better predictions.

Implemented accuracy by splitting the data in training and testing data. I then improved the accuracy by expanding the training options and increasing the amount of epochs from 20 to 32. 32 consistently gave me 90-100% accuracy.

Made a virtual piano. The piano can be controlled with clicks and keyboard inputs.

Implemented the virtual piano with the trained model. When it detects a pointing hand, it will cause a click on wherever the index finger is pointing. You can use this to play the piano with your fingers detected from your webcam.